What I've done since the first commit, which is quite a lot for one single commit. This may be considered the first import of the code. I'll make sure to commit more often..
The module is now able to detect icon sets (a folder with a .info file) in these parts of the Drupal installation:
/icons/*/
/modules/*/icons/
/profiles/*/icons/*/
/sites/*/icons/*/
/sites/*/modules/*/icons/
/sites/*/themes/*/icons/
/themes/*/icons/

Discovered icon sets are registered in the 'system' table. They are also listed with screenshots on admin/build/icons, where the user may choose which icon sets to enable.